Redis mutex, pool hash check null, hard fork 2
Redis mutex - the problem was we had multiple blocks being submitted at the same height creating a lot of orphan blocks. We implemented a semi-working mutex using Redis where the heights of blocks were checked prior to calling "submitblock".

Pool hash check null -  we had a problem where the pool hash would be calculated as null or zero because one of the miner's hashrate was stored in redis as null or zero. We fixed this by first checking if the miner's hashrate is null or zero before being added to the pool's cumulative hashrate

hard fork 2 - we had to decrease the network difficulty by a factor of 16^10 or 1,099,511,627,776 to prevent integer overflow in the Cash2 C++ code. Therefore, any netowrk difficulty values that the pool obtains from a Cash2 node must be multiplied with 16^10 to get the correct value.