Config File : Moved the Cash2 configuration file to the main folder and deleted the config_examples folder because configurations for other coins is not needed
Reason for change : It is easier for people to have just 1 configuration file to work with that is located in the main directory where it belongs. Also, due to the Cash2 mining pool using the Stratum protocol, this mining pool is not usuable with all other coins in the config_examples folder, and thus the folder is no longer needed.