Config File : Renamed the configuration file from config.json to cash2.json
Reason for changes : The gitignore file prevents files with the name config.json from being uploaded to github due to security concerns. Thus, for the configuration file to exist in the main directory, the name had to be change to something other than config.json. So cash2.json was chosen as the name.